UI update > Domain layer folder refactor
