+++ lib/Character_Wiki/Domain/Character/Entity/character.dart
