+++ lib/Character_Wiki/Domain/Character/Repostories/character_repository_impl.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:character_wiki/Character_Wiki/Data/Character/DataSource/character_remote_data_source.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Character/character.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Character/character_repository.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/Entity/character.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/Repostories/character_repository.dart';
 
 class CharacterRepositoryImpl implements CharacterRepository {
   final CharacterRemoteDataSource remoteDataSource;
