+++ lib/Character_Wiki/Domain/Episode/Entity/epsiode.dart
