+++ lib/Character_Wiki/Domain/Episode/Repositories/episode_repository_impl.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:character_wiki/Character_Wiki/Data/Episode/DataSource/episode_remote_data_source.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Episode/epsiode.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Episode/epsiode_repository.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Entity/epsiode.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Repositories/epsiode_repository.dart';
 
 class EpisodeRepositoryImpl implements EpsiodeRepository {
   EpisodeRemoteDataSource episodeRemoteDataSource;
