+++ lib/Character_Wiki/Domain/Episode/Repositories/epsiode_repository.dart
@@ -1,0 +1,5 @@
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Entity/epsiode.dart';
+
+abstract class EpsiodeRepository {
+  Future<List<Episode>> getEpisodes ();
+}
