+++ lib/Character_Wiki/Domain/Episode/UseCases/get_episodes.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:character_wiki/Character_Wiki/Domain/Episode/epsiode.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Episode/epsiode_repository.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Entity/epsiode.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Repositories/epsiode_repository.dart';
 
 class GetEpisodes {
   final EpsiodeRepository episodeRepository;
