--- lib/Character_Wiki/Domain/Episode/epsiode_repository.dart
@@ -1,5 +1,0 @@
-import 'package:character_wiki/Character_Wiki/Domain/Episode/epsiode.dart';
-
-abstract class EpsiodeRepository {
-  Future<List<Episode>> getEpisodes ();
-}
