+++ lib/Character_Wiki/Domain/Location/Entity/location.dart
