+++ lib/Character_Wiki/Domain/Location/Repositories/location_repository.dart
@@ -1,0 +1,5 @@
+import 'package:character_wiki/Character_Wiki/Domain/Location/Entity/location.dart';
+
+abstract class LocationRepository {
+  Future<List<Location>> getLocations ();
+}
