--- lib/Character_Wiki/Domain/Location/location_repository.dart
@@ -1,5 +1,0 @@
-import 'package:character_wiki/Character_Wiki/Domain/Location/location.dart';
-
-abstract class LocationRepository {
-  Future<List<Location>> getLocations ();
-}
