+++ lib/Character_Wiki/Presentation/Bloc/Episode/episode_bloc.dart
@@ -2,8 +2,8 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 
-import 'package:character_wiki/Character_Wiki/Domain/Episode/epsiode.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Episode/get_episodes.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Entity/epsiode.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/UseCases/get_episodes.dart';
 
 part 'episode_event.dart';
 part 'episode_state.dart';
