+++ lib/Character_Wiki/Presentation/Bloc/Location/location_bloc.dart
@@ -2,8 +2,8 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 
-import 'package:character_wiki/Character_Wiki/Domain/Location/get_locations.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Location/location.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Location/UseCases/get_locations.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Location/Entity/location.dart';
 
 part 'location_event.dart';
 part 'location_state.dart';
