+++ lib/service_locator.dart
@@ -1,15 +1,15 @@
 import 'package:character_wiki/Character_Wiki/Data/Character/DataSource/character_remote_data_source.dart';
 import 'package:character_wiki/Character_Wiki/Data/Episode/DataSource/episode_remote_data_source.dart';
 import 'package:character_wiki/Character_Wiki/Data/Location/DataSource/location_remote_data_souce.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Character/character_repository.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Character/character_repository_impl.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Character/get_characters.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Episode/episode_repository_impl.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Episode/epsiode_repository.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Episode/get_episodes.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Location/get_locations.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Location/location_repository.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Location/location_repository_impl.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/Repostories/character_repository.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/Repostories/character_repository_impl.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/UseCase/get_characters.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Repositories/episode_repository_impl.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Repositories/epsiode_repository.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/UseCases/get_episodes.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Location/UseCases/get_locations.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Location/Repositories/location_repository.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Location/Repositories/location_repository_impl.dart';
 import 'package:character_wiki/Character_Wiki/Presentation/Bloc/Character/character_bloc.dart';
 import 'package:character_wiki/Character_Wiki/Presentation/Bloc/Episode/episode_bloc.dart';
 import 'package:character_wiki/Character_Wiki/Presentation/Bloc/Location/location_bloc.dart';
